// flyteconsole/packages/zapp/console/src/components/Launch/LaunchForm/types.ts
export enum InputType {
  Binary = 'BINARY',
  Blob = 'BLOB',
  Boolean = 'BOOLEAN',
  Collection = 'COLLECTION',
  Datetime = 'DATETIME',
  Duration = 'DURATION',
  Error = 'ERROR',
  Enum = 'ENUM',
  Float = 'FLOAT',
  Integer = 'INTEGER',
  Map = 'MAP',
  None = 'NONE',
  Schema = 'SCHEMA',
  String = 'STRING',
  Struct = 'STRUCT',
  Union = 'Union',
  Unknown = 'UNKNOWN',
}

export interface InputTypeDefinition {
  literalType: LiteralType;
  type: InputType;
  subtype?: InputTypeDefinition;
  listOfSubTypes?: InputTypeDefinition[];
}

export interface BlobValue {
  dimensionality: BlobDimensionality | string;
  format?: string;
  uri: string;
}

export interface UnionValue {
  value: InputValue;
  typeDefinition: InputTypeDefinition;
}

export interface NoneValue {}

export type InputValue = string | number | boolean | Date | BlobValue | UnionValue | NoneValue;
export type InputChangeHandler = (newValue: InputValue) => void;

export interface InputProps {
  description: string;
  error?: string;
  helperText?: string;
  initialValue?: Core.ILiteral;
  name: string;
  label: string;
  required: boolean;
  typeDefinition: InputTypeDefinition;
  value?: InputValue;
  onChange: InputChangeHandler;
  setIsError: (boolean) => void;
}